requirements.txt
# Web & API
fastapi
uvicorn[standard]
streamlit
requests
python-multipart  # required by FastAPI for UploadFile / File(...)

# Database
sqlalchemy[asyncio]
asyncpg
pgvector

# AI & Processing
openai
pdfplumber
python-dotenv
pydantic